Add files via upload
This is the C++ code for the 2048 game developed by Mayank Gupta, Manu Jain and Vikas Nyamati, SRM University B. Tech I.T. The code was developed in 2015 during our second semester. The full code has been developed in module and is integrated together in a single CPP file. The code runs fine on The TurboC++ compiler and does support the CodeBlocks compiler due to the intense use of getch() function in the game which is not supported by CodeBlocks.